Delivery

+++ src/en/inter/TrashCan.hx
@@ -21,10 +21,31 @@ class TrashCan extends en.Interactive {
 		ALL.remove(this);
 	}
 
+	function canBeTrashed(k:Data.ItemKind) {
+		return switch( k ) {
+			case Data.ItemKind.Trash, Data.ItemKind.Shit : true;
+			case Data.ItemKind.Fish, FoodBox : false;
+		}
+	}
+
+	override public function canBeActivated(by:Hero) {
+		return super.canBeActivated(by) && by.item!=null;
+	}
+
 	override public function onActivate(by:Hero) {
 		super.onActivate(by);
-		if( by.item!=null )
+		if( by.item!=null && canBeTrashed(by.item) )
 			by.destroyItem();
+	}
+
+	override function onTouch(e:Entity) {
+		super.onTouch(e);
+		if( e.is(en.inter.ItemDrop) ) {
+			var e = e.as(en.inter.ItemDrop);
+			if( canBeTrashed(e.k) )
+				e.destroy();
+		}
+
 	}
 
 	override public function postUpdate() {
